added a new footer to main tableview
